third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-send-sharedarraybuffer-expected.txt

CONSOLE WARNING: line 12: SharedArrayBuffer will require cross-origin isolation as of M92, around July 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
This is a testharness.js-based test.
PASS sending SharedArrayBuffer
Harness: the test ran to completion.
